move some files to ./libs
