+++ main.html
@@ -9,8 +9,8 @@
     <!-- Prevent caching at the proxy server -->
     <meta http-equiv="expires" content="0">
     <link rel="stylesheet" type="text/css" href="main.css" media="screen"/>
-    <script type="text/javascript" src="jquery-3.3.1.js"></script>
-    <script type="text/javascript" src="md5.js"></script>
+    <script type="text/javascript" src="libs/jquery-3.3.1.js"></script>
+    <script type="text/javascript" src="libs/md5.js"></script>
     <script type="text/javascript" src="proto.js"></script>
     <script type="module" src="main.js"></script>
   </head>
@@ -47,7 +47,7 @@
           </div>
         </div>
       </div>
-    </div>    
+    </div>
     <div class="dlg-cover dlg-alert" style="display:none;">
       <div class="dlg">
         <div class="dlg-box">
