+++ manifest.json
@@ -7,24 +7,24 @@
     "icons": {
 	"64": "icons/bee.png"
     },
-    "content_security_policy": "script-src 'self'; object-src 'self'; navigation-to 'filesystem:'" ,
-    "permissions": [	
-        "nativeMessaging", 
-        "contextMenus",
-        "menus",
-        "downloads",
-        "bookmarks",
-        "storage",
-        "unlimitedStorage",
+    "content_security_policy": "script-src 'self'; object-src 'self'; navigation-to 'filesystem:'",
+    "permissions": [
         "activeTab",
-        "tabs",
-        "webNavigation",
-        "webRequest",
-        "webRequestBlocking",
+        "bookmarks",
+        "contextMenus",
+        "downloads",
         "file:///*",
         "http://*/*",
         "https://*/*",
-	"notifications"
+        "menus",
+        "nativeMessaging",
+        "notifications",
+        "storage",
+        "tabs",
+        "unlimitedStorage",
+        "webNavigation",
+        "webRequest",
+        "webRequestBlocking"
     ],
     "applications": {
         "gecko": {
@@ -52,14 +52,14 @@
                 "http://*/*",
                 "https://*/*"
             ],
-            "js": [
-                "jquery-3.3.1.js",
-                "proto.js",
-                "md5.js",
-                "content_script.js"
-            ],
             "css": [
                 "content_script.css"
+            ],
+            "js": [
+                "libs/jquery-3.3.1.js",
+                "libs/md5.js",
+                "proto.js",
+                "content_script.js"
             ]
         }
     ],
