+++ options.html
@@ -4,7 +4,7 @@
     <title>ScrapBee -> {Setting}</title>
     <link rel="icon" href="icons/bee.svg" type="image/png" />
     <link rel="stylesheet" type="text/css" href="options.css" media="all"/>
-    <script type="text/javascript" src="jquery-3.3.1.js"></script>
+    <script type="text/javascript" src="libs/jquery-3.3.1.js"></script>
     <script type="text/javascript" src="proto.js"></script>
     <script type="module" src="options.js"></script>
   </head>
