+++ search.html
@@ -3,7 +3,7 @@
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
     <title>Scrapbee -> {Search}</title>
     <link rel="icon" href="icons/bee.svg" type="image/png" />
-    <script type="text/javascript" src="jquery-3.3.1.js"></script>
+    <script type="text/javascript" src="libs/jquery-3.3.1.js"></script>
     <script type="module" src="proto.js"></script>
     <script type="module" src="search.js"></script>
     <style type="text/css" media="screen">
